refactor retypes

+++ contracts/governance/extensions/GovernorPreventLateQuorum.sol
@@ -23,7 +23,7 @@ abstract contract GovernorPreventLateQuorum is Governor {
     uint64 private _voteExtension;
 
     /// @custom:oz-retyped-from mapping(uint256 => Timers.BlockNumber)
-    mapping(uint256 => uint256) private _extendedDeadlines;
+    mapping(uint256 => uint64) private _extendedDeadlines;
 
     /// @dev Emitted when a proposal deadline is pushed back due to reaching quorum late in its voting period.
     event ProposalExtended(uint256 indexed proposalId, uint64 extendedDeadline);
