+++ contracts/governance/extensions/GovernorTimelockCompound.sol
@@ -26,7 +26,7 @@ abstract contract GovernorTimelockCompound is IGovernorTimelock, Governor {
     ICompoundTimelock private _timelock;
 
     /// @custom:oz-retyped-from mapping(uint256 => Timers.Timestamp)
-    mapping(uint256 => uint256) private _proposalTimelocks;
+    mapping(uint256 => uint64) private _proposalTimelocks;
 
     /**
      * @dev Emitted when the timelock controller used for proposal execution is modified.
